requirements.txt
pip-tools
ipython
ipdb
flake8
Django
psycopg2
raven
WebTest
django-webtest
model-mommy
django-debug-toolbar
django-model-utils
django-redis
hiredis
django-bootstrap3
django-vanilla-views
django-extensions
django-clear-cache
uwsgi
django-configurations
whitenoise
# brotlipy
invoke
pip-tools
cookiecutter
flake8
pytest
sh